+++ herostratus/src/rules/mod.rs
@@ -3,6 +3,7 @@ mod h001_fixup;
 mod h002_shortest_subject_line;
 mod h003_longest_subject_line;
 mod h004_non_unicode;
+mod h005_empty_commit;
 #[cfg(test)]
 pub(crate) mod test_rules;